#95 time keeper unsigned long. vrfd
almost done
todo cleanups
todo doxygen

+++ src/apps/clock/clockApp.cpp
@@ -140,7 +140,7 @@ static unsigned int getStepCounter()
     return (bmaApi->getCounter)();
 }
 
-static unsigned int getNextWakeUpPeriod()
+static unsigned long getNextWakeUpPeriod()
 {
     return NW_DONT_CARE;
 }
